+++ src/vp8/common/arm/neon/idct_blk_neon.rs
@@ -27,8 +27,8 @@ pub unsafe fn vp8_dequant_idct_add_y_block_neon(
     let mut i: i32 = 0;
     i = 0 as i32;
     while i < 4 as i32 {
-        if *(eobs as *mut i16).offset(0 as i32 as isize) != 0 {
-            if *(eobs as *mut i16).offset(0 as i32 as isize)
+        if *(eobs as *mut i16).offset(0 as isize) != 0 {
+            if *(eobs as *mut i16).offset(0 as isize)
                 as i32
                 & 0xfefe as i32
                 != 0
@@ -37,36 +37,36 @@ pub unsafe fn vp8_dequant_idct_add_y_block_neon(
             } else {
                 idct_dequant_0_2x_neon(
                     q as *mut int16_t,
-                    *dq.offset(0 as i32 as isize) as int16_t,
+                    *dq.offset(0 as isize) as int16_t,
                     dst,
                     stride,
                 );
             }
         }
-        if *(eobs as *mut i16).offset(1 as i32 as isize) != 0 {
-            if *(eobs as *mut i16).offset(1 as i32 as isize)
+        if *(eobs as *mut i16).offset(1 as isize) != 0 {
+            if *(eobs as *mut i16).offset(1 as isize)
                 as i32
                 & 0xfefe as i32
                 != 0
             {
                 idct_dequant_full_2x_neon(
-                    q.offset(32 as i32 as isize),
+                    q.offset(32 as isize),
                     dq as *mut int16_t,
-                    dst.offset(8 as i32 as isize),
+                    dst.offset(8 as isize),
                     stride,
                 );
             } else {
                 idct_dequant_0_2x_neon(
-                    q.offset(32 as i32 as isize),
-                    *dq.offset(0 as i32 as isize) as int16_t,
-                    dst.offset(8 as i32 as isize),
+                    q.offset(32 as isize),
+                    *dq.offset(0 as isize) as int16_t,
+                    dst.offset(8 as isize),
                     stride,
                 );
             }
         }
-        q = q.offset(64 as i32 as isize);
+        q = q.offset(64 as isize);
         dst = dst.offset((4 as i32 * stride) as isize);
-        eobs = eobs.offset(4 as i32 as isize);
+        eobs = eobs.offset(4 as isize);
         i += 1;
     }
 }
@@ -79,8 +79,8 @@ pub unsafe fn vp8_dequant_idct_add_uv_block_neon(
     mut stride: i32,
     mut eobs: *mut i8,
 ) {
-    if *(eobs as *mut i16).offset(0 as i32 as isize) != 0 {
-        if *(eobs as *mut i16).offset(0 as i32 as isize)
+    if *(eobs as *mut i16).offset(0 as isize) != 0 {
+        if *(eobs as *mut i16).offset(0 as isize)
             as i32
             & 0xfefe as i32
             != 0
@@ -89,16 +89,16 @@ pub unsafe fn vp8_dequant_idct_add_uv_block_neon(
         } else {
             idct_dequant_0_2x_neon(
                 q as *mut int16_t,
-                *dq.offset(0 as i32 as isize) as int16_t,
+                *dq.offset(0 as isize) as int16_t,
                 dst_u,
                 stride,
             );
         }
     }
-    q = q.offset(32 as i32 as isize);
+    q = q.offset(32 as isize);
     dst_u = dst_u.offset((4 as i32 * stride) as isize);
-    if *(eobs as *mut i16).offset(1 as i32 as isize) != 0 {
-        if *(eobs as *mut i16).offset(1 as i32 as isize)
+    if *(eobs as *mut i16).offset(1 as isize) != 0 {
+        if *(eobs as *mut i16).offset(1 as isize)
             as i32
             & 0xfefe as i32
             != 0
@@ -107,15 +107,15 @@ pub unsafe fn vp8_dequant_idct_add_uv_block_neon(
         } else {
             idct_dequant_0_2x_neon(
                 q as *mut int16_t,
-                *dq.offset(0 as i32 as isize) as int16_t,
+                *dq.offset(0 as isize) as int16_t,
                 dst_u,
                 stride,
             );
         }
     }
-    q = q.offset(32 as i32 as isize);
-    if *(eobs as *mut i16).offset(2 as i32 as isize) != 0 {
-        if *(eobs as *mut i16).offset(2 as i32 as isize)
+    q = q.offset(32 as isize);
+    if *(eobs as *mut i16).offset(2 as isize) != 0 {
+        if *(eobs as *mut i16).offset(2 as isize)
             as i32
             & 0xfefe as i32
             != 0
@@ -124,16 +124,16 @@ pub unsafe fn vp8_dequant_idct_add_uv_block_neon(
         } else {
             idct_dequant_0_2x_neon(
                 q as *mut int16_t,
-                *dq.offset(0 as i32 as isize) as int16_t,
+                *dq.offset(0 as isize) as int16_t,
                 dst_v,
                 stride,
             );
         }
     }
-    q = q.offset(32 as i32 as isize);
+    q = q.offset(32 as isize);
     dst_v = dst_v.offset((4 as i32 * stride) as isize);
-    if *(eobs as *mut i16).offset(3 as i32 as isize) != 0 {
-        if *(eobs as *mut i16).offset(3 as i32 as isize)
+    if *(eobs as *mut i16).offset(3 as isize) != 0 {
+        if *(eobs as *mut i16).offset(3 as isize)
             as i32
             & 0xfefe as i32
             != 0
@@ -142,7 +142,7 @@ pub unsafe fn vp8_dequant_idct_add_uv_block_neon(
         } else {
             idct_dequant_0_2x_neon(
                 q as *mut int16_t,
-                *dq.offset(0 as i32 as isize) as int16_t,
+                *dq.offset(0 as isize) as int16_t,
                 dst_v,
                 stride,
             );
